(wip) Starting to clean up mess
From previous checkin.

A lot of the code that I rolled into here relies on libraries that
either weren't/aren't part of the project or have been deprecated in
favor of others.

This was basically just an exercise in letting the compiler find these
errors for me.

Not completed, by any means. But it's forward progress.